lyfRest/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.11
Django==4.0.2
django-cors-headers==3.10.1
djangorestframework==3.13.1
gcloud==0.18.3
googleapis-common-protos==1.54.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
jws==0.1.3
oauth2client==4.1.3
Pillow==8.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.14.1
pyparsing==3.0.7
Pyrebase4==4.5.0
python-dotenv==0.19.2
python-jwt==2.0.1
pytz==2021.3
reportlab==3.6.5
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
